requirements.txt
uvicorn
fastapi
yake==0.4.8
pydantic_settings
kafka-python
